+++ src/Domain/Repository/UserRepositoryInterface.php
@@ -15,6 +15,7 @@ use Proximum\Vimeet\Domain\Model\Event;
 use Proximum\Vimeet\Domain\Model\PaginatedResult;
 use Proximum\Vimeet\Domain\Model\Product;
 use Proximum\Vimeet\Domain\Model\Sheet;
+use Proximum\Vimeet\Domain\Model\Type;
 use Proximum\Vimeet\Domain\Model\Unavailability\Mass;
 use Proximum\Vimeet\Domain\Model\User;
 
@@ -104,10 +105,11 @@ interface UserRepositoryInterface
     /**
      * @param Event  $event
      * @param string $locale
+     * @param Type[] $types
      *
      * @return UserSheetTypeView[]
      */
-    public function getWithSheetAndTypeByEvent(Event $event, string $locale): array;
+    public function getWithSheetAndTypeByEvent(Event $event, string $locale, array $types): array;
 
     /**
      * @param Event $event