+++ src/Ui/Bundle/AdminBundle/Controller/Rooming/RoomingListAction.php
@@ -5,6 +5,9 @@ namespace Proximum\Vimeet\Ui\Bundle\AdminBundle\Controller\Rooming;
 use Proximum\Vimeet\Application\Adapter\QueryBusInterface;
 use Proximum\Vimeet\Application\Query\Rooming\RoomingList\ListViewQuery;
 use Proximum\Vimeet\Domain\Model\Event;
+use Proximum\Vimeet\Ui\Bundle\AdminBundle\Form\Type\Rooming\FilterType;
+use Proximum\Vimeet\Ui\Bundle\AdminBundle\ValueResolver\AdminDomain;
+use Symfony\Component\Form\FormFactoryInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Templating\EngineInterface;
@@ -17,20 +20,33 @@ class RoomingListAction
     /** @var QueryBusInterface */
     private $queryBus;
 
-    public function __construct(EngineInterface $engine, QueryBusInterface $queryBus)
+    /** @var FormFactoryInterface */
+    private $formFactory;
+
+    public function __construct(EngineInterface $engine, QueryBusInterface $queryBus, FormFactoryInterface $formFactory)
     {
         $this->engine = $engine;
         $this->queryBus = $queryBus;
+        $this->formFactory = $formFactory;
     }
 
-    public function __invoke(Request $request, Event $event): Response
+    public function __invoke(Request $request, Event $event, AdminDomain $adminDomain): Response
     {
         $locale = $event->getAvailableLocale($request->getLocale());
-        $roomingListView = $this->queryBus->handle(new ListViewQuery($event, $locale));
+        $listViewQuery = new ListViewQuery($event, $locale, []);
+
+        $form = $this->formFactory->create(FilterType::class, $listViewQuery, [
+            'event' => $event, 'locale' => $locale, 'admin' => $adminDomain->getAdmin(),
+        ]);
+
+        $form->handleRequest($request);
+
+        $roomingListView = $this->queryBus->handle($listViewQuery);
 
         return new Response($this->engine->render('@Admin/Rooming/RoomingList/list.html.twig', [
             'event' => $event,
             'roomingListView' => $roomingListView,
+            'form' => $form->createView(),
         ]));
     }
 }